Refactor move DeleteConversationDialog tests
